+++ shims/spark30/src/main/scala/com/nvidia/spark/rapids/shims/spark30/GpuBroadcastNestedLoopJoinExec.scala
@@ -16,24 +16,24 @@
 
 package com.nvidia.spark.rapids.shims.spark30
 
-import com.nvidia.spark.rapids._
-import org.apache.spark.sql.rapids.execution._
+import com.nvidia.spark.rapids.{GpuBuildLeft, GpuBuildRight, GpuBuildSide}
 
-import org.apache.spark.internal.Logging
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.plans.JoinType
 import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.sql.execution.joins.{BroadcastNestedLoopJoinExec, BuildLeft, BuildRight}
+import org.apache.spark.sql.rapids.execution._
 
-
+/**
+ *  Spark 3.1 changed packages of BuildLeft, BuildRight, BuildSide
+ */
 case class GpuBroadcastNestedLoopJoinExec(
     left: SparkPlan,
     right: SparkPlan,
     join: BroadcastNestedLoopJoinExec,
     joinType: JoinType,
     condition: Option[Expression])
-  extends GpuBroadcastNestedLoopJoinExecBase(left, right, join, joinType, condition) with Logging {
-
+  extends GpuBroadcastNestedLoopJoinExecBase(left, right, join, joinType, condition) {
 
   def getBuildSide: GpuBuildSide = {
     join.buildSide match {
@@ -42,18 +42,4 @@ case class GpuBroadcastNestedLoopJoinExec(
       case _ => throw new Exception("unknown buildSide Type")
     }
   }
-}
-
-object GpuBroadcastNestedLoopJoinExec extends Logging {
-
-  def createInstance(
-      left: SparkPlan,
-      right: SparkPlan,
-      join: BroadcastNestedLoopJoinExec,
-      joinType: JoinType,
-      condition: Option[Expression]): GpuBroadcastNestedLoopJoinExecBase= {
-    
-    GpuBroadcastNestedLoopJoinExec(left, right, join, joinType, condition)
-  }
-
 }
