+++ shims/spark30/src/main/scala/com/nvidia/spark/rapids/shims/spark30/GpuFirst.scala
@@ -16,14 +16,16 @@
 
 package com.nvidia.spark.rapids.shims.spark30
 
-import org.apache.spark.sql.rapids.GpuFirstBase
-import com.nvidia.spark.rapids._
+import com.nvidia.spark.rapids.GpuLiteral
 
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult.{TypeCheckFailure, TypeCheckSuccess}
 import org.apache.spark.sql.catalyst.expressions.{Expression, Literal}
+import org.apache.spark.sql.rapids.GpuFirstBase
 
-
+/**
+ * Parameters to GpuFirst changed in Spark 3.1
+ */
 case class GpuFirst(child: Expression, ignoreNullsExpr: Expression) extends GpuFirstBase(child) {
   override def children: Seq[Expression] = child :: ignoreNullsExpr :: Nil
