+++ shims/spark31/src/main/scala/com/nvidia/spark/rapids/shims/spark31/GpuLast.scala
@@ -16,13 +16,14 @@
 
 package com.nvidia.spark.rapids.shims.spark31
 
+import org.apache.spark.sql.catalyst.analysis.TypeCheckResult
+import org.apache.spark.sql.catalyst.analysis.TypeCheckResult.TypeCheckSuccess
+import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.rapids.GpuLastBase
 
-import org.apache.spark.sql.catalyst.analysis.TypeCheckResult
-import org.apache.spark.sql.catalyst.analysis.TypeCheckResult.{TypeCheckFailure, TypeCheckSuccess}
-import org.apache.spark.sql.catalyst.expressions.Expression
-
-
+/**
+ * Parameters to GpuFirst changed in Spark 3.1
+ */
 case class GpuLast(child: Expression, ignoreNulls: Boolean) extends GpuLastBase(child) {
   override def children: Seq[Expression] = child :: Nil
