+++ shims/spark31/src/main/scala/com/nvidia/spark/rapids/shims/spark31/Spark31Shims.scala
@@ -19,21 +19,19 @@ package com.nvidia.spark.rapids.shims.spark31
 import java.time.ZoneId
 
 import com.nvidia.spark.rapids._
-import com.nvidia.spark.rapids.shims.spark31.GpuBroadcastNestedLoopJoinExec
 
 import org.apache.spark.sql.catalyst.expressions.aggregate.{First, Last}
-import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, BuildSide}
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.execution._
 import org.apache.spark.sql.execution.datasources.HadoopFsRelation
-import org.apache.spark.sql.execution.joins.{BroadcastHashJoinExec, BroadcastNestedLoopJoinExec, SortMergeJoinExec}
+import org.apache.spark.sql.execution.joins.{BroadcastHashJoinExec, BroadcastNestedLoopJoinExec, HashJoin, SortMergeJoinExec}
 import org.apache.spark.sql.execution.joins.ShuffledHashJoinExec
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.plans.JoinType
 import org.apache.spark.sql.rapids.GpuTimeSub
 import org.apache.spark.sql.rapids.execution.GpuBroadcastNestedLoopJoinExecBase
 import org.apache.spark.sql.rapids.shims.spark31._
-import org.apache.spark.unsafe.types.{CalendarInterval, UTF8String}
+import org.apache.spark.unsafe.types.CalendarInterval
 import org.apache.spark.sql.types._
 
 class Spark31Shims extends SparkShims with Logging {
@@ -135,41 +133,11 @@ class Spark31Shims extends SparkShims with Logging {
     )
   }
 
-  def getBuildSide(join: ShuffledHashJoinExec): GpuBuildSide = {
-    val buildSide = join.buildSide
-    buildSide match {
-      case e: buildSide.type if e.toString.contains("BuildRight") => {
-        GpuBuildRight
-      }
-      case l: buildSide.type if l.toString.contains("BuildLeft") => {
-        GpuBuildLeft
-      }
-      case _ => throw new Exception("unknown buildSide Type")
-    }
+  def getBuildSide(join: HashJoin): GpuBuildSide = {
+    GpuJoinUtils.getBuildSide(join.buildSide)
   }
 
   def getBuildSide(join: BroadcastNestedLoopJoinExec): GpuBuildSide = {
-    val buildSide = join.buildSide
-    buildSide match {
-      case e: buildSide.type if e.toString.contains("BuildRight") => {
-        GpuBuildRight
-      }
-      case l: buildSide.type if l.toString.contains("BuildLeft") => {
-        GpuBuildLeft
-      }
-      case _ => throw new Exception("unknown buildSide Type")
-    }
-  }
-  def getBuildSide(join: BroadcastHashJoinExec): GpuBuildSide = {
-    val buildSide = join.buildSide
-    buildSide match {
-      case e: buildSide.type if e.toString.contains("BuildRight") => {
-        GpuBuildRight
-      }
-      case l: buildSide.type if l.toString.contains("BuildLeft") => {
-        GpuBuildLeft
-      }
-      case _ => throw new Exception("unknown buildSide Type")
-    }
+    GpuJoinUtils.getBuildSide(join.buildSide)
   }
 }
