+++ shims/spark31/src/main/scala/org/apache/spark/sql/rapids/shims/spark31/GpuFileSourceScanExec.scala
@@ -19,7 +19,6 @@ package org.apache.spark.sql.rapids.shims.spark31
 import java.util.concurrent.TimeUnit.NANOSECONDS
 
 import com.nvidia.spark.rapids.{GpuExec, GpuReadCSVFileFormat, GpuReadOrcFileFormat, GpuReadParquetFileFormat, SparkPlanMeta}
-import org.apache.spark.sql.rapids.GpuFileSourceScanExecBase
 
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.{InternalRow, TableIdentifier}
@@ -31,6 +30,7 @@ import org.apache.spark.sql.execution.datasources.csv.CSVFileFormat
 import org.apache.spark.sql.execution.datasources.orc.OrcFileFormat
 import org.apache.spark.sql.execution.datasources.parquet.ParquetFileFormat
 import org.apache.spark.sql.execution.metric.SQLMetric
+import org.apache.spark.sql.rapids.GpuFileSourceScanExecBase
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.vectorized.ColumnarBatch
 import org.apache.spark.util.collection.BitSet
